Added skeleton for pettitt test
Updated the pettitt test function so that most elements from the VB code are represented in some way. More work needs to be done, mostly ensuring that all processes are producing their desired result and also finding some way to condense the may for loops that appear at the end of the function.